src/main/java/com/twu/outputwriter/ConsoleOutputWriter.java
package com.twu.outputwriter;

import com.twu.menu.Output;

import java.io.BufferedWriter;
import java.io.IOException;

public class ConsoleOutputWriter implements OutputWriter {
    private final BufferedWriter bufferedWriter;

    public ConsoleOutputWriter(BufferedWriter bufferedWriter) {
        this.bufferedWriter = bufferedWriter;
    }

    @Override
    public void write(Output outputMessages) {
        try {
            for(String output : outputMessages.getOutput())
                bufferedWriter.write(output + "\n");
            bufferedWriter.flush();
        } catch (IOException e) {
            e.printStackTrace();
        }
    }
}
